+++ src/components/dashboard/manage-account/TwoFAModal.tsx
@@ -1,12 +1,12 @@
 'use client';
 import { useEffect, useState } from "react";
-import { AlertDialog, AlertDialogContent, AlertDialogHeader, AlertDialogTitle, AlertDialogTrigger } from "./ui/alert-dialog";
-import { Button } from "./ui/button";
+import { AlertDialog, AlertDialogContent, AlertDialogHeader, AlertDialogTitle, AlertDialogTrigger } from "@/components/ui/alert-dialog";
+import { Button } from "@/components/ui/button";
 import { Loader, X } from "lucide-react";
 import { z } from "zod";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
-import { Input } from "./ui/input";
+import { Input } from "@/components/ui/input";
 import toast from "react-hot-toast";
 
 const formSchema = z.object({